+++ src/Mutation/MutationGenerator.php
@@ -33,7 +33,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\Mutant\Generator;
+namespace Infection\Mutation;
 
 use function assert;
 use function count;
@@ -61,7 +61,7 @@ use Throwable;
 /**
  * @internal
  */
-final class MutationsGenerator
+final class MutationGenerator
 {
     /**
      * @var SplFileInfo[]
@@ -118,9 +118,7 @@ final class MutationsGenerator
         $this->eventDispatcher->dispatch(new MutationGeneratingStarted(count($this->sourceFiles)));
 
         foreach ($this->sourceFiles as $file) {
-            if (!$onlyCovered || $this->hasTests($file)) {
-                $allFilesMutations[] = $this->getMutationsFromFile($file, $onlyCovered, $extraNodeVisitors);
-            }
+            $allFilesMutations[] = $this->getMutationsFromFile($file, $onlyCovered, $extraNodeVisitors);
 
             $this->eventDispatcher->dispatch(new MutableFileProcessed());
         }
@@ -138,6 +136,10 @@ final class MutationsGenerator
      */
     private function getMutationsFromFile(SplFileInfo $file, bool $onlyCovered, array $extraNodeVisitors): array
     {
+        if (!$onlyCovered || $this->hasTests($file)) {
+            return [];
+        }
+
         try {
             /** @var Node[] $initialStatements */
             $initialStatements = $this->parser->parse($file->getContents());
