+++ tests/phpunit/Mutation/MutationGeneratorTest.php
@@ -33,7 +33,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Mutant\Generator;
+namespace Infection\Tests\Mutation;
 
 use Infection\EventDispatcher\EventDispatcherInterface;
 use Infection\Events\MutableFileProcessed;
@@ -42,7 +42,7 @@ use Infection\Events\MutationGeneratingStarted;
 use Infection\Exception\InvalidMutatorException;
 use Infection\FileSystem\SourceFileCollector;
 use Infection\Mutant\Exception\ParserException;
-use Infection\Mutant\Generator\MutationsGenerator;
+use Infection\Mutation\MutationGenerator;
 use Infection\Mutator\Arithmetic\Decrement;
 use Infection\Mutator\Arithmetic\Plus;
 use Infection\Mutator\Boolean\TrueValue;
@@ -58,7 +58,7 @@ use PhpParser\ParserFactory;
 use PHPUnit\Framework\TestCase;
 use Pimple\Container;
 
-final class MutationsGeneratorTest extends TestCase
+final class MutationGeneratorTest extends TestCase
 {
     private const FIXTURES_DIR = __DIR__ . '/../../Fixtures/Files';
 
@@ -170,7 +170,7 @@ final class MutationsGeneratorTest extends TestCase
                 [new MutationGeneratingFinished()]
             );
 
-        $generator = new MutationsGenerator(
+        $generator = new MutationGenerator(
             (new SourceFileCollector())->collectFiles(
                 [self::FIXTURES_DIR . '/Mutation/OneFile'],
                 [],
@@ -190,7 +190,7 @@ final class MutationsGeneratorTest extends TestCase
         ?string $whitelistedMutatorName = null,
         ?MutatorConfig $mutatorConfig = null,
         array $srcDirs = []
-    ): MutationsGenerator {
+    ): MutationGenerator {
         if ($srcDirs === []) {
             $srcDirs = [
                 self::FIXTURES_DIR . '/Mutation/OneFile',
@@ -230,7 +230,7 @@ final class MutationsGeneratorTest extends TestCase
         $eventDispatcherMock = $this->createMock(EventDispatcherInterface::class);
         $eventDispatcherMock->expects($this->any())->method('dispatch');
 
-        return new MutationsGenerator(
+        return new MutationGenerator(
             (new SourceFileCollector())->collectFiles(
                 $srcDirs,
                 $excludedDirsOrFiles,
